refactor(drawer): extract file editor into its own component
Also moved state tracking of edited longURl into LongUrlEditor
component instead of redux
